Created README. Renamed the large simulation main file to large_simulation.cpp.
