+++ Project-2/Problem-1b/README.md
@@ -1,0 +1,25 @@
+# Problem2b implementation
+
+Problem 2b was about implementing the necessary classes for the Restricted Boltzmann Machine type of wave function, plus updating the gradient descent and start experimenting with learning rate. 
+
+# Building the code
+If run from the root directory of the Git-repo, the build commands are:
+
+```
+make app=Project-2/Problem-1b/unit_tests
+make app=Project-2/Problem-1b/single_particle 
+make app=Project-2/Problem-1b/large_simulation
+```
+
+# Running tests
+
+To run the unit tests, run this program:
+```
+bin/Project-2/Problem-1b/unit_tests.out
+```
+
+No parameter needed. It runs some tests, print various output, and if everything is ok ends with the line: All tests passed!
+
+# Running simulations
+
+TDB
